requirements.txt
# EEG Mental State Classifier — Python Dependencies

# Core scientific stack
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Deep Learning
tensorflow>=2.13.0

# EEG-specific
mne>=1.4.0          # EDF reading, preprocessing, visualization
wfdb>=4.1.0         # PhysioNet data download

# Signal processing
PyWavelets>=1.4.1   # Wavelet transforms (optional feature)

# Config
PyYAML>=6.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.0.0
